Get the RTC working
This required some PCB redesigns

I couldn't get the RTC to work on other pins than the defaults (4, 5) so I had to
move the LED pin over to pin 13 since it used to be on 5.  This was done by hacking
the PCB, but could be made permanent in a future board if I end up making more.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>