Get All the LEDs working
There seems to be an out-of-memory issue going on tho, I was only able to get it
to work with all the LEDs by removing 64 bytes from a string buffer (for the clock
string for debugging).  That _works_ but it's kinda pathetic and means I'm *riggght*
on the edge of running out of meory, which is no good since this is just a
proof-of-concept FW at this point...

I'm not sure if FastLED or the RTC library is using most of it up or what's going
on, but I hope I can figure it out because there's still a lot of functionality
that I want to add and I'm probably going to need a few more bytes of memory to
do it.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>