Increase the buffer size for the clock -- hacky fix
I'm not sure what the deal is honestly, but it seems like there's some
garbage going on with my Clock :(  It's clobbering my code somehow
I added a bunch of extra space after it and it makes everything work.

A terrible terrible solution, hacking as heck but it's a step in the
right direction.

Signed-off-by: Charlie Mooney <cmooney3@gmail.com>